# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEBMASTERY SCHOOL API DOCS",
        "description": "Webmastery School Portal API documentation",
        "contact": {
            "email": "narushevich.maksim@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7000/api/v1",
            "description": "API documentation"
        }
    ],
    "paths": {
...
        "/sections": {
            "get": {
                "tags": [
                    "Section"
                ],
                "summary": "Get all sections",
                "description": "Returns all sections details",
                "operationId": "getSections",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Section page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items to be retrieved per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "relationships",
                        "in": "query",
                        "description": "Relationships that should be inclused in response",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "course"
                    },
                    {
                        "name": "relationPerPage",
                        "in": "query",
                        "description": "Number of relationships to be retrieved in response structure",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received available sections",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Section"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Section"
                ],
                "summary": "Create a new section",
                "description": "Create a new section.",
                "operationId": "createSection",
                "requestBody": {
                    "description": "Create section",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSection"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Section created"
                    },
                    "400": {
                        "description": "Request validation error"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sections/{id}": {
            "get": {
                "tags": [
                    "Section"
                ],
                "summary": "Get a specific section",
                "description": "Returns section details",
                "operationId": "getSectionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Section id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received specific sections",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Section"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "404": {
                        "description": "Section Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Section"
                ],
                "summary": "Delete section",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteSection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of sections to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Section deleted"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "404": {
                        "description": "Section not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Section"
                ],
                "summary": "Update a specific section",
                "description": "Update section.",
                "operationId": "updateSection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Section id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update section",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section updated"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "400": {
                        "description": "Request validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Section": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "course_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CreateSection": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Section"
                    },
                    {
                        "required": [
                            "name",
                            "description",
                            "course_id"
                        ]
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Oauth2 security",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}